fix: UI changes
