+++ code-mentor-ai/src/app/page.tsx
@@ -5,7 +5,7 @@ export default function Home() {
   return (
     <div className="dark:bg-[#050505] bg-white min-h-screen">
       <Navbar />
-      <div className="pt-48 px-4 sm:px-18 md:px-32">
+      <div className="pt-48 px-6 s:px-24 sm:px-18 md:px-32">
 
         <div className="p-2 lg:pr-[8vw] xl:pr-[15vw] text-2xl md:text-3xl lg:text-4xl font-bold text-gray-800 dark:text-white">
           Smarter Code Reviews, Powered by AI
@@ -18,7 +18,8 @@ export default function Home() {
         <div className="pl-2">
           <Link href="/code-review">
             <button className="border border-[#00000069] dark:border-[#fff6] dark:text-white 
-            p-2 px-4 rounded hover:bg-gray-300 hover:border-gray-300 dark:hover:bg-[#2d2d2d] dark:hover:border-[#2d2d2d]">
+            p-2 px-4 rounded hover:bg-gray-500 hover:text-white hover:border-gray-300 dark:hover:bg-[#2d2d2d] 
+            dark:hover:border-[#2d2d2d]">
               Review Code
             </button>
           </Link>
